requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
paddlepaddle==2.4.0
paddleocr==2.6.0.1
pillow==9.5.0
numpy==1.24.3